refator: create core module and separate user entity & repository from auth module
