+++ docker-compose.yml
@@ -1,6 +1,6 @@
 services:
   db:
-    image: postgres:16-alpine
+    image: public.ecr.aws/docker/library/postgres:16-alpine
     container_name: app_db
     environment:
       POSTGRES_USER: app
@@ -13,6 +13,7 @@ services:
   app:
     build: { context: . }
     container_name: app_backend
+    user : root
     working_dir: /app
     command: sh -c "chmod +x ./gradlew && ./gradlew bootRun --no-daemon"
     volumes:
