+++ src/main/java/com/coffeeteam/fitbyte/auth/service/AuthServiceImpl.java
@@ -2,17 +2,15 @@ package com.coffeeteam.fitbyte.auth.service;
 
 import com.coffeeteam.fitbyte.auth.dto.UserAuthRequest;
 import com.coffeeteam.fitbyte.auth.dto.UserAuthResponse;
-import com.coffeeteam.fitbyte.auth.entity.User;
-import com.coffeeteam.fitbyte.auth.repository.UserRepository;
+import com.coffeeteam.fitbyte.core.entity.User;
+import com.coffeeteam.fitbyte.core.repository.UserRepository;
 import com.coffeeteam.fitbyte.auth.security.JwtUtil;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
-import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.web.server.ResponseStatusException;
 
-import java.time.LocalDateTime;
 import java.time.ZonedDateTime;
 import java.util.Optional;
