+++ src/main/java/com/coffeeteam/fitbyte/core/entity/User.java
@@ -1,11 +1,10 @@
-package com.coffeeteam.fitbyte.auth.entity;
+package com.coffeeteam.fitbyte.core.entity;
 
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.time.LocalDateTime;
 import java.time.ZonedDateTime;
 
 @Data
