+++ src/main/java/com/coffeeteam/fitbyte/core/repository/UserRepository.java
@@ -1,6 +1,6 @@
-package com.coffeeteam.fitbyte.auth.repository;
+package com.coffeeteam.fitbyte.core.repository;
 
-import com.coffeeteam.fitbyte.auth.entity.User;
+import com.coffeeteam.fitbyte.core.entity.User;
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
 import org.springframework.data.jpa.repository.JpaRepository;
